requirements.txt
appdirs==1.4.4
asgiref==3.5.0
audioread==2.1.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.6
numba==0.55.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
pooch==1.6.0
pycparser==2.21
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytube==12.0.0
pytz==2022.1
requests==2.27.1
resampy==0.2.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SoundFile==0.10.3.post1
spotipy==2.19.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
urllib3==1.26.9
ytmusicapi==0.21.0
zipp==3.8.0
